include cli option for torch's built in matmul precision
When using a supported card and using preview build of pytorch (currently tested on version 2.8.0.dev20250616+cu128 and NVIDIA 5070 Ti) pytorch informs about the availability of matmulprecision which results in drastically improved runtimes when using high or medium instead of highest setting. This commit includes a command line option to toggle this based on user preference, default is highest. Keeping the default at highest should not cause any compatibility issues, as this is also the current default.