Only consider standard residues in consecutive CA filter
The consecutive CA filter marks a protein chain as invalid if it has two consecutive residues that are too far apart. I recently found a CIF file where a few zinc and calcium ions were included in the protein chain. Since the ions are far apart, this chain would previously have been marked invalid by the filter. After this small change, the ions are ignored by the filter and the protein chain is marked as valid during processing.